requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
python-multipart==0.0.6
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gunicorn>=21.0.0
Brotli>=1.1.0
notion-client>=2.2.0
python-dotenv>=1.0.0
python-telegram-bot>=20.7 